プロパティ生成関数を別ファイルに切り出した
